fix(web): preserve origin main navigation changes
